experimental alternative StateGenerator

+++ app/src/main/java/de/felixnuesse/timedsilence/fragments/graph/GraphItemView.kt
@@ -51,7 +51,7 @@ class GraphItemView: LinearLayout {
 
     }
 
-    fun updateVisibility(duration: Int) {
+    fun updateVisibility(duration: Long) {
         if(duration<10) {
             binding.textContainer.visibility = View.GONE
         }
